# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Smart Archive API - أرشيف ذكي
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sa/notifications/{nid}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: nid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationReadInput"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /sa/projects/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /sa/projects/{id}/documents/{did}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: did
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /sa/voice:
    post:
      operationId: processVoice
      tags: [voice]
      summary: Process a voice command (text extracted from audio)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceInput"
      responses:
        "200":
          description: Voice processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceResult"
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [reminder, info, warning, success]
        priority:
          type: string
          enum: [high, medium, low]
        scheduledAt:
          type: ["string", "null"]
        read:
          type: boolean
        projectId:
          type: ["string", "null"]
        actionUrl:
          type: ["string", "null"]
          description: Frontend path to navigate to when notification is clicked
        createdByName:
          type: ["string", "null"]
          description: Display name of the user who triggered this notification
        createdAt:
          type: string
      required: [id, title, message, type, read, createdAt]
    NotificationReadInput:
      type: object
      properties:
        read:
          type: boolean
      required: [read]
...
    Document:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [pdf, image, word, excel, other]
        url:
          type: string
        size:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, projectId, name, type, url, createdAt]
    DocumentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [pdf, image, word, excel, other]
        url:
          type: string
        size:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [name, type, url]
...
    VoiceInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        projectId:
          type: ["string", "null"]
      required: [text]
    VoiceResult:
      type: object
      properties:
        action:
          type: string
          enum: [reminder, search, add, unknown]
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [action, message, data]